backend/requirements.txt
Django==3.1.4
djangorestframework==3.12.4
django-cors-headers
dj-database-url==0.5.0
PyJWT==2.0.1
gunicorn==20.0.4
psycopg2==2.8.5
psycopg2-binary==2.8.5
whitenoise==5.2.0
python-dotenv